+++ js/generatePhotoGallery/main.js
@@ -1,0 +1,3 @@
+import {generatePhotoGallery} from './data.js';
+
+generatePhotoGallery(25);
